bazel: add dev roachprod-bench command
Add command roachprod-bench to dev. The command will build all the test binaries
for a given set of packages, using Bazel. Any runfiles required by the binaries
will also be staged. Binaries and runfiles are added to a tar in artifacts and
can be optionally compressed. The output is portable without being dependent on
the repository.

Each package is supplied with a run script that will setup environment variables
for Bazel runfiles to be used by the test binary.

Additionally, this builds the binary for a sub command roachprod-bench that will
orchestrate the execution of the binaries. The sub command is also executed by
the roachprod-bench command with specified arguments passed to it.

Resolves: #90837
See also: #90958

Epic: CRDB-20903

Release justification: test-only change.
Release note: None