rpc: share loopback connections between rpc contexts
Currently, we supply rpc context with a loopback connection function when it is
not required to go through the tcp stack. However, when we instantiate another
rpc context, as is the case with in-process multi-tenant, that rpc context will
not be aware of the loopback connection present in the process.

To share these loopback connections for different ports between multiple rpc
contexts a global var is added to maintain a map of addresses to loopback
dialers. When dialing, the loopback map will be checked first to see if a local
transport (vs. tcp transport) can be used.

A shutdown hook is also added to remove any addresses from the loopback map when
stopping.