roachprod: fix dns create records
Previously, `CreateRecords` could either update existing records, or create new
records. The cache captured only new records and discarded existing records.
This is incorrect as the cache should match the records that get set by the CLI
command, which is the combination of the existing and new records for a name
entry.

This change updates the code so that the cache entry is also updated with the
combined set.

Epic: None
Release Note: None