roachtest: add resize operation
This change adds a new resize operation that will grow clusters and start
cockroach on the new nodes based on the existing cockroach start script on the
first node of the cluster. It will copy the certs, binary and libs from the
first node as well.

In the new state the operation will pause for a duration before draining and
then decommissioning the new nodes.

If the operation succeeded the cleanup function will shrink the cluster back to
its original size. But if the operation fails, the cleanup function is not
invoked and the cluster will remain at its new size. This is useful for
investigating the new nodes for issues, but also should be monitored since if
the operation is run again it will keep increasing the cluster in size.

Release Note: None
Epic: None