roachtest: interface for goroutines
Previously, roachtests would use bare goroutines `go func...` to start tasks
during a test to perform various background functions. This is problematic due
to the nature of roachtest, which runs multiple tests in parallel using workers.
If one such bare goroutine panics the whole test suite is interrupted and the
process exits.

This change is the first step and introduces an interface that both the
traditional and mixedversion test frameworks can utilise. It intends to target
both regular and MVT tests by supplying the test with the interface either
through a helper or the test interface.

In addition to providing the `Go` function, there are additional options that
can be passed. The test framework will generally supply the panic and error
handler, but users can supply additional options such as naming the task, or
passing a custom logger. This makes logging more useful to determine what task
panic or encountered an error.

Informs: #118214

Epic: None
Release note: None