roachtest: task manager
This change provides a task manager implementation that will drive the logic
running behind the Tasker interface, as well as provide the implementation of
it. The manager is meant to be used by the test framework(s), while the Tasker
interface is supplied to tests. The framework will do additional logic such as
terminating tasks at the end of a test, logging uncaught errors or issues and
providing the appropriate panic and error handlers.

Informs: #118214

Epic: None
Release note: None