roachtest: add gomock target for Test interface
Previously, any updates to the test interface would cause the build to fail
since the mock for it in the `clusterstats` package was not automatically
generated. This change adds it to the bazel build in order to regenerate the
mock for the roachtest Test interface.

Epic: None
Release note: None